+++ include/activation/sigmoid.hpp
@@ -10,11 +10,11 @@ void
 Sigmoid::init(int cur_in_size,
               int cur_input_row,
               int cur_input_col,
-              double node_num,
-              double kernel_row,
-              double kernel_col,
-              double stride_row,
-              double stride_col,
+              int node_num,
+              int kernel_row,
+              int kernel_col,
+              int stride_row,
+              int stride_col,
               std::string padding,
               std::string name) {
     m_type = Layer::Sigmoid;
