+++ include/layers/Dense.hpp
@@ -19,11 +19,11 @@ public:
     void init(int cur_in_size,
               int cur_input_row,
               int cur_input_col,
-              double node_num,
-              double kernel_row,
-              double kernel_col,
-              double stride_row,
-              double stride_col,
+              int node_num,
+              int kernel_row,
+              int kernel_col,
+              int stride_row,
+              int stride_col,
               std::string padding,
               std::string name);  ///< Initialize the layer with those arguments
     void init(Eigen::MatrixXd weight, Eigen::MatrixXd bias);   ///< Initialize the layer with given kernel and bias
